refactor: split settings_screen.dart into separate files
