+++ lib/screens/8_settings/pin_dialog.dart
@@ -1,0 +1,194 @@
+import 'package:flutter/material.dart';
+import 'package:lucide_icons/lucide_icons.dart';
+import '../../core/constants/colors.dart';
+
+/// Diálogo para ingresar PIN de 4 dígitos
+class PinDialog extends StatefulWidget {
+  final String title;
+  final String subtitle;
+
+  const PinDialog({
+    super.key,
+    required this.title,
+    required this.subtitle,
+  });
+
+  @override
+  State<PinDialog> createState() => _PinDialogState();
+}
+
+class _PinDialogState extends State<PinDialog> {
+  final List<String> _pin = [];
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: AppColors.deepVoidPurple,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
+      title: Text(
+        widget.title,
+        style: const TextStyle(
+          fontFamily: 'Inter',
+          fontSize: 20,
+          fontWeight: FontWeight.bold,
+          color: Colors.white,
+        ),
+        textAlign: TextAlign.center,
+      ),
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Text(
+            widget.subtitle,
+            style: TextStyle(
+              fontFamily: 'Inter',
+              fontSize: 14,
+              color: AppColors.textSecondary,
+            ),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 24),
+          // Indicadores de PIN
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: List.generate(4, (index) {
+              final filled = index < _pin.length;
+              return Container(
+                margin: const EdgeInsets.symmetric(horizontal: 8),
+                width: 16,
+                height: 16,
+                decoration: BoxDecoration(
+                  shape: BoxShape.circle,
+                  color: filled
+                      ? AppColors.primaryAction
+                      : Colors.white.withValues(alpha: 0.2),
+                  border: Border.all(
+                    color: filled
+                        ? AppColors.primaryAction
+                        : Colors.white.withValues(alpha: 0.3),
+                    width: 2,
+                  ),
+                ),
+              );
+            }),
+          ),
+          const SizedBox(height: 24),
+          // Teclado numérico
+          _buildNumPad(),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildNumPad() {
+    return Column(
+      children: [
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: ['1', '2', '3'].map((n) => _buildNumKey(n)).toList(),
+        ),
+        const SizedBox(height: 12),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: ['4', '5', '6'].map((n) => _buildNumKey(n)).toList(),
+        ),
+        const SizedBox(height: 12),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: ['7', '8', '9'].map((n) => _buildNumKey(n)).toList(),
+        ),
+        const SizedBox(height: 12),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          children: [
+            // Cancelar
+            GestureDetector(
+              onTap: () => Navigator.pop(context),
+              child: Container(
+                width: 60,
+                height: 60,
+                decoration: BoxDecoration(
+                  color: Colors.white.withValues(alpha: 0.05),
+                  borderRadius: BorderRadius.circular(30),
+                ),
+                child: Icon(
+                  LucideIcons.x,
+                  color: AppColors.textSecondary,
+                  size: 24,
+                ),
+              ),
+            ),
+            _buildNumKey('0'),
+            // Borrar
+            GestureDetector(
+              onTap: _deleteDigit,
+              child: Container(
+                width: 60,
+                height: 60,
+                decoration: BoxDecoration(
+                  color: Colors.white.withValues(alpha: 0.05),
+                  borderRadius: BorderRadius.circular(30),
+                ),
+                child: Icon(
+                  LucideIcons.delete,
+                  color: AppColors.textSecondary,
+                  size: 24,
+                ),
+              ),
+            ),
+          ],
+        ),
+      ],
+    );
+  }
+
+  Widget _buildNumKey(String number) {
+    return GestureDetector(
+      onTap: () => _addDigit(number),
+      child: Container(
+        width: 60,
+        height: 60,
+        decoration: BoxDecoration(
+          color: Colors.white.withValues(alpha: 0.1),
+          borderRadius: BorderRadius.circular(30),
+        ),
+        child: Center(
+          child: Text(
+            number,
+            style: const TextStyle(
+              fontFamily: 'Inter',
+              fontSize: 24,
+              fontWeight: FontWeight.w600,
+              color: Colors.white,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+
+  void _addDigit(String digit) {
+    if (_pin.length < 4) {
+      setState(() {
+        _pin.add(digit);
+      });
+
+      if (_pin.length == 4) {
+        // PIN completo, cerrar con resultado
+        Future.delayed(const Duration(milliseconds: 200), () {
+          if (mounted) {
+            Navigator.pop(context, _pin.join());
+          }
+        });
+      }
+    }
+  }
+
+  void _deleteDigit() {
+    if (_pin.isNotEmpty) {
+      setState(() {
+        _pin.removeLast();
+      });
+    }
+  }
+}
